==> 03_get_taxi_data.ipynb <==
import json
import os
import pandas as pd
import requests
from datetime import datetime as dt
from dateutil.relativedelta import relativedelta

#Ez csak 1 utazást ad vissza
url = 'https://data.cityofchicago.org/resource/wrvz-psew.json?trip_id=00000019ef3c0b419a9e6b1843b83530b79627a4'
headers = {"X-App-Token": os.environ.get("CHICAGO_API_TOKEN")}
response = requests.get(url, headers)
data = response.json()

print(os.environ.get("CHICAGO_API_TOKEN"))

headers = {"X-App-Token": os.environ.get("CHICAGO_API_TOKEN")}
response = requests.get(url, headers)
data = response.json()

headers


api_token = os.environ.get("CHICAGO_API_TOKEN")
url = f'https://data.cityofchicago.org/resource/wrvz-psew.json?trip_id=00000019ef3c0b419a9e6b1843b83530b79627a4&$$app_token={api_token}'
response = requests.get(url)
data = response.json()


data

# Egy napi adat
# Az URL-be SoQL-t használ a lekérdezés összeállításához https://dev.socrata.com/docs/queries/
url = f"https://data.cityofchicago.org/resource/wrvz-psew.json?$where=trip_start_timestamp >= '2023-08-05T00:00:00' AND trip_start_timestamp <= '2023-08-05T23:59:59'&$limit=30000&$$app_token={api_token}"
headers = {"X-App-Token": os.environ.get("CHICAGO_API_TOKEN")}
response = requests.get(url)
data = response.json()
response

url = f"https://data.cityofchicago.org/resource/wrvz-psew.json?$where=trip_start_timestamp >= '2023-08-05T00:00:00' AND trip_start_timestamp <= '2023-08-05T23:59:59'&$limit=30000&$$app_token={api_token}"
#headers = {"X-App-Token": os.environ.get("CHICAGO_API_TOKEN")}
response = requests.get(url)
data = response.json()
response

#Mivel nekünk a T-2 hónap kell ezért módosítani kell a dátumot
# Egy napi adat
# Az URL-be SoQL-t használ a lekérdezés összeállításához https://dev.socrata.com/docs/queries/

startDate = dt.now().strftime("%Y-%m-%d")
endDate = (dt.now() - relativedelta(months=2)).strftime("%Y-%m-%d")

startDate = '2023-09-05'

url = f"https://data.cityofchicago.org/resource/wrvz-psew.json?$where=trip_start_timestamp >= '{startDate}T00:00:00' AND trip_start_timestamp <= '{startDate}T23:59:59'&$limit=30000"
headers = {"X-App-Token": os.environ.get("CHICAGO_API_TOKEN")}
response = requests.get(url, headers)
data = response.json()
response

len(data)

data
