==> 07-transform_load.ipynb <==
import pandas as pd
import requests
import os
from dateutil.relativedelta import relativedelta
from datetime import datetime as dt
import json

pd.set_option('display.max_columns',30)

"""
Steps to do

1. Get the data from S3
2. weather data transformations
3. taxi trips transformation - Done
4. Update payment type - Done
5. Update company - Done
6. update taxi trips with company and paymant types (replace str values with ids from the latest master tables) - Done
7. upload the weather to S3
8. upload the taxi data to S3
9. upload the newest payment type and company

"""

#### Taxi trips transformation code

start_date = (dt.now() - relativedelta(months=2)).strftime("%Y-%m-%d")
url_new = f"https://data.cityofchicago.org/resource/ajtu-isnz.json?$where=trip_start_timestamp >= '{start_date}T00:00:00' AND trip_start_timestamp <= '{start_date}T23:59:59'&$limit=213000000"
response_taxi = requests.get(url_new)
taxi_data = response_taxi.json()

taxi_trips = pd.DataFrame(taxi_data)

taxi_trips.sample(100)

taxi_trips.dtypes

taxi_trips.drop(['pickup_census_tract', 'dropoff_census_tract'], axis=1, inplace=True)
taxi_trips.drop(['pickup_centroid_location', 'dropoff_centroid_location'], axis=1, inplace=True)
taxi_trips.dropna(inplace=True)


# taxi_trips['trip_start_timestamp'] = taxi_trips['trip_start_timestamp'].astype('datetime64[ms]')
# taxi_trips['trip_end_timestamp'] = taxi_trips['trip_end_timestamp'].astype('datetime64[ms]')
# taxi_trips['trip_seconds'] = taxi_trips['trip_seconds'].astype('int64')
# #taxi_trips['trip_miles'] = taxi_trips['trip_miles'].astype('float64')
# taxi_trips['pickup_community_area'] = taxi_trips['pickup_community_area'].astype('int64')
# taxi_trips['dropoff_community_area'] = taxi_trips['dropoff_community_area'].astype('int64')
# #taxi_trips['fare'] = taxi_trips['fare'].astype('float64')
# # taxi_trips['tips'] = taxi_trips['tips'].astype('float64')
# # taxi_trips['tolls'] = taxi_trips['tolls'].astype('float64')
# # taxi_trips['extras'] = taxi_trips['extras'].astype('float64')
# #taxi_trips['trip_total'] = taxi_trips['trip_total'].astype('float64')



taxi_trips.to_csv('taxi_trips.csv', index=False, sep = ';', encoding = 'utf-16')

taxi_trips.dtypes

taxi_trips.rename(columns={'pickup_community_area' : 'pickup_community_area_id',
                           'dropoff_community_area' : 'dropoff_community_area_id'
                           },inplace=True)

taxi_trips['datetime_for_weather'] = pd.to_datetime(taxi_trips['trip_start_timestamp']).dt.floor('H')
#taxi_trips['datetime_for_weather'] = taxi_trips['trip_start_timestamp'].dt.floor('H')

taxi_trips.head()

#### Taxi trip transformation function

def taxi_trips_transformations(taxi_trips: pd.DataFrame) -> pd.DataFrame:
    """
    Transformations for taxi trips DataFrame.

    Parameters:
    - taxi_trips (pd.DataFrame): DataFrame containing taxi trips data.

    Returns:
    - pd.DataFrame: Transformed DataFrame with specified columns dropped, NaN values removed,
      columns renamed, and datetime rounded to the nearest hour for weather matching.
    """
    if not isinstance(taxi_trips,pd.DataFrame):
      raise TypeError('taxi_trips is not a valid data frame')
    
    taxi_trips.drop(['pickup_census_tract', 'dropoff_census_tract','pickup_centroid_location', 'dropoff_centroid_location'], axis=1, inplace=True)
    taxi_trips.dropna(inplace=True)
    taxi_trips.rename(columns={'pickup_community_area' : 'pickup_community_area_id',
                            'dropoff_community_area' : 'dropoff_community_area_id'
                            },inplace=True)
    taxi_trips['datetime_for_weather'] = pd.to_datetime(taxi_trips['trip_start_timestamp']).dt.floor('H')
    return taxi_trips

#### Company update code

company_master = taxi_trips['company'].drop_duplicates().reset_index(drop=True)
company_master = pd.DataFrame({
    'company_id': range(1, len(company_master) + 1),
    'company':  company_master
})

company_master.tail()

new_company_data = [
    {'company' : 'Petani Cab Corp'},
    {'company' : 'X'},
    {'company' : 'Y'}
]

new_company_mapping = pd.DataFrame(new_company_data)

new_company_mapping

company_max_id = company_master['company_id'].max()

new_companies = []
for company in new_company_mapping['company'].values:
    if company not in company_master['company'].values:
        new_companies.append(company)

#One line
new_companies_on_line = [company for company in new_company_mapping['company'].values if company not in company_master['company'].values]

new_companies_df = pd.DataFrame({
    'company_id' : range(company_max_id + 1, company_max_id + len(new_companies_on_line) + 1),
    'company' : new_companies_on_line
})

new_companies_df.info()

updated_company_master = pd.concat([company_master, new_companies_df], ignore_index=True)

updated_company_master.tail()

#### Create company master update function


def update_company_master(taxi_trips : pd.DataFrame, company_master: pd.DataFrame) -> pd.DataFrame:
    """
    Update the company master dataframe with new companies from the taxi trips dataframe.
    
    Args:
        taxi_trips (pd.DataFrame): The dataframe containing taxi trips data.
        company_master (pd.DataFrame): The dataframe containing the current company master data.
    
    Returns:
        pd.DataFrame: The updated company master dataframe with new companies added.
    """
    company_max_id = company_master['company_id'].max()

    new_companies = [company for company in taxi_trips['company'].values if company not in company_master['company'].values]
    
    new_companies_df = pd.DataFrame({
        'company_id' : range(company_max_id + 1, company_max_id + len(new_companies) + 1),
        'company' : new_companies
    })

    return pd.concat([company_master, new_companies_df], ignore_index=True)

taxi_trips_company_only_for_test = pd.DataFrame({
    'company_id': [1,2,3],
    'company': ['5167 - 71969 5167 Taxi Inc','X','Y']
})

test_df = update_company_master(taxi_trips=taxi_trips_company_only_for_test, company_master=company_master)

test_df.tail(3)

#### Create payment type update function

#Ha új payment type van azt frissíteni kell
payment_type_master = taxi_trips['payment_type'].drop_duplicates().reset_index(drop=True)
payment_type_master = pd.DataFrame({
    'payment_type_id': range(1, len(payment_type_master) + 1),
    'payment_type':  payment_type_master
})

taxi_trips_payment_only_for_test = pd.DataFrame({
    'payment_type_id': [1,2,3],
    'payment_type': ['Credit Card','F','G']
})

taxi_trips_payment_only_for_test

def update_payment_master(taxi_trips : pd.DataFrame, payment_type_master: pd.DataFrame) -> pd.DataFrame:
    """Update the payment master with new payment types from taxi trips.
    
    Args:
        taxi_trips (pd.DataFrame): DataFrame containing taxi trips data.
        payment_type_master (pd.DataFrame): DataFrame containing existing payment types.
    
    Returns:
        pd.DataFrame: Updated payment master DataFrame.
    """
    
    payment_max_id = payment_type_master['payment_type_id'].max()

    new_payments = [payments for payments in taxi_trips['payment_type'].values if payments not in payment_type_master['payment_type'].values]
    
    new_payments_df = pd.DataFrame({
        'payment_type_id' : range(payment_max_id + 1, payment_max_id + len(new_payments) + 1),
        'payment_type' : new_payments
    })

    return pd.concat([payment_type_master, new_payments_df], ignore_index=True)

test_df = update_payment_master(taxi_trips=taxi_trips_payment_only_for_test, payment_type_master=payment_type_master)

test_df

#### Unify the two update function

#A cél, hogy ne duplikáld a kódot.
def update_master(taxi_trips : pd.DataFrame, master: pd.DataFrame, id_col_name:str, value_col_name:str) -> pd.DataFrame:
    """Update the master dataframe with new values from the taxi trips dataframe.
    
    Args:
        taxi_trips (pd.DataFrame): The dataframe containing new values.
        master (pd.DataFrame): The master dataframe to be updated.
        id_col_name (str): The name of the column in the master dataframe that contains IDs.
        value_col_name (str): The name of the column in the master dataframe that contains values.
    
    Returns:
        pd.DataFrame: The updated master dataframe with new values added.
    """
    
    max_id = master[id_col_name].max()

    new_values = [value for value in taxi_trips[value_col_name].values if value not in master[value_col_name].values]
    
    new_data_df = pd.DataFrame({
        id_col_name : range(max_id + 1, max_id + len(new_values) + 1),
        value_col_name : new_values
    })

    return pd.concat([master, new_data_df], ignore_index=True)

test_df = update_master(taxi_trips=taxi_trips_payment_only_for_test, master=payment_type_master, id_col_name='payment_type_id', value_col_name='payment_type')

test_df

test_df = update_master(taxi_trips=taxi_trips_company_only_for_test, master=company_master, id_col_name='company_id', value_col_name='company')

test_df

#### Update taxi_trips with the most recent company_master and payment_type codes

#A nevek lecserélése az id-re, ez a cella hozzáfűzi a kódokat az eredeti df-hez.
#Ezek után törölni lehet az eredeti mezőket (payment_type, company)
taxi_trips_id = taxi_trips.merge(payment_type_master, on='payment_type')
taxi_trips_id = taxi_trips_id.merge(company_master, on='company')


taxi_trips_id.head(5)

taxi_trips_id.drop(['payment_type', 'company'], axis=1, inplace=True)

taxi_trips_id.sample(5)

def update_taxi_trips_with_master_data(taxi_trips: pd.DataFrame, payment_type_master: pd.DataFrame, company_master: pd.DataFrame) -> pd.DataFrame:
    """Update taxi trips with master data.
    
    Parameters:
    taxi_trips (pd.DataFrame): DataFrame containing taxi trips data.
    payment_type_master (pd.DataFrame): DataFrame containing master data for payment types.
    company_master (pd.DataFrame): DataFrame containing master data for companies.
    
    Returns:
    pd.DataFrame: Updated DataFrame with merged master data.
    """
        
    taxi_trips_id = taxi_trips.merge(payment_type_master, on='payment_type')
    taxi_trips_id = taxi_trips_id.merge(company_master, on='company')
    taxi_trips_id.drop(['payment_type', 'company'], axis=1, inplace=True)
    return taxi_trips_id

test_df = update_taxi_trips_with_master_data(taxi_trips=taxi_trips, payment_type_master=payment_type_master, company_master=company_master)
test_df.sample(5)

#### Weather transformation function

def transform_weather_data(weather_data: json) -> pd.DataFrame:
    """Transform weather data from JSON format to a pandas DataFrame.
    
    Parameters:
    weather_data (json): JSON data containing weather information.
    
    Returns:
    pd.DataFrame: Transformed weather data in a DataFrame format.
    """
    
    weather_filtered = {
        'datetime': weather_data['hourly']['time'],
        'temperature' : weather_data['hourly']['temperature_2m'],
        'wind' : weather_data['hourly']['wind_speed_10m'],
        'precipitation' : weather_data['hourly']['precipitation'],
        'rain' : weather_data['hourly']['rain']
    }
    weather_df = pd.DataFrame(weather_filtered)
    weather_df['datetime'] = pd.to_datetime(weather_df['datetime'])
    return weather_df


#Test
url = f"https://archive-api.open-meteo.com/v1/era5"
startDate = (dt.now() - relativedelta(months=2)).strftime('%Y-%m-%d')

params = {
    "latitude" : 41.85,
    "longitude" : -87.65,
    "start_date" : startDate,
    "end_date" : startDate,
    "hourly" : "temperature_2m,wind_speed_10m,precipitation,rain"
}


response = requests.get(url, params=params)
weather_data = response.json()

weather_data_df = transform_weather_data(weather_data=weather_data)

weather_data_df.sample(5)
