==> 08-local_visualization.ipynb <==
import os
import pandas as pd
import boto3
from io import StringIO
from dotenv import load_dotenv

pd.set_option('display.max_columns', 50)

load_dotenv()
aws_access_key = os.getenv('AWS_ACCESS_KEY')
aws_secret_key = os.getenv('AWS_SECRET_KEY')

def read_csv_from_s3(bucket: str, path: str, filename: str) -> pd.DataFrame:
    """
    Reads a CSV file from an Amazon S3 bucket into a Pandas DataFrame.

    Parameters:
    - bucket (str): The name of the Amazon S3 bucket.
    - path (str): The path within the bucket where the CSV file is located.
    - filename (str): The name of the CSV file to be read.

    Returns:
    - pd.DataFrame: A Pandas DataFrame containing the data read from the CSV file.

    Note:
    - Requires the boto3 library to be installed.
    - Assumes the CSV file is UTF-8 encoded.
    """
    
    s3 = boto3.client('s3', aws_access_key_id = aws_access_key, aws_secret_access_key = aws_secret_key)
    full_path = f'{path}{filename}'
    
    object = s3.get_object(Bucket=bucket, Key=full_path)
    object = object['Body'].read().decode('utf-8')
    
    return pd.read_csv(StringIO(object))

bucket = 'chicago-taxi'
payment_type_path = 'transformed_data/payment_type/'
payment_type_file = 'payment_type_master.csv'

community_areas_path = 'transformed_data/community_areas/'
community_areas_file = 'community_areas_master.csv'

company_path = 'transformed_data/company/'
company_file = 'company_master.csv'

date_path = 'transformed_data/date/'
date_file = 'date_dim.csv'

weather_path = 'transformed_data/weather/'
weather_file = ''

taxi_trips_path = 'transformed_data/taxi_trips/'
taxi_trips_file = ''

s3 = boto3.client('s3', aws_access_key_id = aws_access_key, aws_secret_access_key = aws_secret_key)

community_areas = read_csv_from_s3(bucket=bucket, path=community_areas_path, filename=community_areas_file )
company = read_csv_from_s3(bucket=bucket, path=company_path, filename=company_file )
date = read_csv_from_s3(bucket=bucket, path=date_path, filename=date_file )
payment_type = read_csv_from_s3(bucket=bucket, path=payment_type_path, filename=payment_type_file )

trips_list = []
weather_list = []

#Taxi data transformation and loading
for file in s3.list_objects(Bucket = bucket, Prefix = taxi_trips_path)['Contents']:
    taxi_trip_key = file['Key']
    if taxi_trip_key.split('/')[-1].strip() != '':
        if taxi_trip_key.split('.')[1] == 'csv':
            filename = taxi_trip_key.split('/')[-1].strip()
            trip = read_csv_from_s3(bucket=bucket, path=taxi_trips_path, filename=filename)
            trips_list.append(trip)
            print(f'{filename} has been added.')

trips_df = pd.concat(trips_list, ignore_index=True)

trips_df.head()

#Weather data trandform and loading
for file in s3.list_objects(Bucket = bucket, Prefix = weather_path)['Contents']:
    weather_key = file['Key']
    
    if weather_key.split('/')[-1].strip() != '':
        if weather_key.split('.')[1] == 'csv':
            filename = weather_key.split('/')[-1].strip()
            weather = read_csv_from_s3(bucket=bucket, path=weather_path, filename=filename)
            weather_list.append(weather)
            print(f'{filename} has been added.')


weather_df = pd.concat(weather_list, ignore_index=True)

weather_df.head()

community_areas.head()

company.head()

date.head()

payment_type.head()

#### Join data

trips_full_df = pd.merge(trips_df, weather_df, left_on='datetime_for_weather', right_on='datetime', how='inner')
trips_full_df = trips_full_df.drop(columns='datetime_for_weather')

trips_full_df = pd.merge(trips_full_df, company, left_on='company_id', right_on='company_id', how='inner')
trips_full_df = trips_full_df.drop(columns='company_id')

trips_full_df = pd.merge(trips_full_df, payment_type, left_on='payment_type_id', right_on='payment_type_id', how='inner')
trips_full_df = trips_full_df.drop(columns='payment_type_id')

trips_full_df = pd.merge(trips_full_df, community_areas, left_on='pickup_community_area_id', right_on='Area code', how='inner')
trips_full_df = trips_full_df.drop(columns=['pickup_community_area_id', 'Area code'])

trips_full_df = trips_full_df.rename(columns={'Area name': 'pickup_community_area'})

trips_full_df.head()

trips_full_df = pd.merge(trips_full_df, community_areas, left_on='dropoff_community_area_id', right_on='Area code', how='inner')
trips_full_df = trips_full_df.drop(columns=['dropoff_community_area_id', 'Area code'])
trips_full_df = trips_full_df.rename(columns={'Area name': 'dropoff_community_area'})

date['date'] = pd.to_datetime(date['date'])
date.info()

trips_full_df['trip_start_timestamp'] = pd.to_datetime(trips_full_df['trip_start_timestamp'])

trips_full_df['trip_start_date'] = trips_full_df['trip_start_timestamp'].dt.date
trips_full_df['trip_start_date'] = pd.to_datetime(trips_full_df['trip_start_date'])

trips_full_df = pd.merge(trips_full_df, date, left_on='trip_start_date', right_on='date', how='inner')

trips_full_df.drop(columns=['date'], inplace=True)
trips_full_df.head()
